README.md
# Investegator
Desktop application that store criminals and officers info in xml files using C# and it displays their info and sevral statisics about crimes and criminals. it even suggests criminals to a specific crime according to the evidence found in crime area and the another info about the criminal


This Project has been published after it has been discussed with the TAs so no contribution appears for some of the teammates.